# docs/openapi.json
{
  "definitions": {
...
    "ClientResponse": {
      "properties": {
        "company": {
          "type": "string"
        },
        "contactEmail": {
          "format": "email",
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "extNumber": {
          "type": "integer"
        },
        "id": {
          "type": "string"
        },
        "intNumber": {
          "type": "integer"
        },
        "phone": {
          "type": "string"
        },
        "projects": {
          "items": {
            "$ref": "#/definitions/ProjectResponse"
          },
          "type": "array"
        },
        "rfc": {
          "type": "string"
        },
        "street": {
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        }
      },
      "required": [
        "company",
        "contactEmail",
        "createdAt",
        "extNumber",
        "id",
        "intNumber",
        "phone",
        "projects",
        "rfc",
        "street",
        "updatedAt",
        "zipCode"
      ],
      "type": "object"
    },
    "ClientUpdateRequest": {
      "properties": {
        "company": {
          "type": "string"
        },
        "contactEmail": {
          "format": "email",
          "type": "string"
        },
        "extNumber": {
          "type": "integer"
        },
        "intNumber": {
          "type": "integer"
        },
        "phone": {
          "type": "string"
        },
        "rfc": {
          "type": "string"
        },
        "street": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
...
  "info": {
    "title": "ProfileMatcher",
    "version": "1.0.0"
  },
  "paths": {
...
    "/manager/v1/clients": {
      "get": {
        "description": "Get all clients",
        "parameters": [
          
        ],
        "responses": {
          "200": {
            "description": "The clients found",
            "schema": {
              "items": {
                "$ref": "#/definitions/ClientResponse"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "Client"
        ]
      },
      "post": {
        "description": "Add a client",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ClientPostRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The client was added",
            "schema": {
              "$ref": "#/definitions/ClientResponse"
            }
          }
        },
        "tags": [
          "Client"
        ]
      }
    },
    "/manager/v1/clients/{clientId}": {
      "delete": {
        "description": "Delete a client by id",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "The client was deleted",
            "schema": {
              
            }
          }
        },
        "tags": [
          "Client"
        ]
      },
      "get": {
        "description": "Get a client by id",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The client found",
            "schema": {
              "$ref": "#/definitions/ClientResponse"
            }
          }
        },
        "tags": [
          "Client"
        ]
      },
      "put": {
        "description": "Update a client by id",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ClientUpdateRequest"
            }
          },
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The client updated",
            "schema": {
              "$ref": "#/definitions/ClientResponse"
            }
          }
        },
        "tags": [
          "Client"
        ]
      }
    },
    "/manager/v1/clients/{clientId}/projects": {
      "get": {
        "description": "Get all projects by clientId",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The projects found by specific client",
            "schema": {
              "items": {
                "$ref": "#/definitions/ProjectResponse"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "Client"
        ]
      },
      "post": {
        "description": "Add a project for a specific client",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProjectPostRequest"
            }
          },
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The projects found by specific client",
            "schema": {
              "items": {
                "$ref": "#/definitions/ProjectResponse"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "Client"
        ]
      }
    },
    "/manager/v1/clients/{clientId}/projects/{projectId}": {
      "delete": {
        "description": "Delete a specific project for a specific client",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "The project was deleted",
            "schema": {
              
            }
          }
        },
        "tags": [
          "Client"
        ]
      },
      "get": {
        "description": "Get a specific project for a specific client",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The project found by specific client",
            "schema": {
              "$ref": "#/definitions/ProjectResponse"
            }
          }
        },
        "tags": [
          "Client"
        ]
      },
      "put": {
        "description": "Update a specific project for a specific client",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProjectUpdateRequest"
            }
          },
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The project updated",
            "schema": {
              "items": {
                "$ref": "#/definitions/ProjectResponse"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "Client"
        ]
      }
    },
    "/manager/v1/clients/{clientId}/projects/{projectId}/profiles": {
      "get": {
        "description": "Get all required profiles for a specific project from specific client",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          
        },
        "tags": [
          "Client"
        ]
      },
      "post": {
        "deprecated": true,
        "description": "Add a required profile for a specific project from specific client",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          
        },
        "tags": [
          "Client"
        ]
      }
    },
    "/manager/v1/clients/{clientId}/projects/{projectId}/profiles/{profileId}": {
      "delete": {
        "deprecated": true,
        "description": "Delete a required profile for a specific project from specific client",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "The project's profile was deleted",
            "schema": {
              
            }
          }
        },
        "tags": [
          "Client"
        ]
      },
      "get": {
        "deprecated": true,
        "description": "Get a required profile for a specific project from specific client",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          
        },
        "tags": [
          "Client"
        ]
      },
      "put": {
        "deprecated": true,
        "description": "Update a required profile for a specific project from specific client",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          
        },
        "tags": [
          "Client"
        ]
      }
    },
...
    "/manager/v1/profiles/{profileId}/certifications": {
      "post": {
        "description": "Add certification for a specific profile",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProfileCertificationPostRequest"
            }
          },
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "The certification was added",
            "schema": {
              "$ref": "#/definitions/ProfileCertificationResponse"
            }
          }
        },
        "tags": [
          "Profile"
        ]
      }
    },
...
  },
  "swagger": "2.0.0"
}